AC-514: PatientBirthDateValidatorWatcherTest returns a NullPointerException
