+++ openmrs-client/build.gradle
@@ -41,7 +41,7 @@ android {
         targetSdkVersion 27
 
         testApplicationId "org.openmrs.mobile.test"
-        testInstrumentationRunner "org.openmrs.mobile.test.OpenMRSInstrumentationTestRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
     configurations.all {
@@ -101,6 +101,11 @@ dependencies {
 
     implementation 'com.google.code.gson:gson:2.8.2'
     androidTestImplementation(
+            'junit:junit:4.12',
+            'androidx.test.ext:junit:1.0.0',
+            'androidx.test.espresso:espresso-core:3.1.0',
+            'com.android.support.test:rules:1.0.2',
+            'com.android.support.test:runner:1.0.2',
             'org.mockito:mockito-core:1.10.19',
             'com.google.dexmaker:dexmaker-mockito:1.2',
             'com.google.dexmaker:dexmaker:1.2',
